requirements.txt
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
docopt==0.6.2
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
Flask==3.0.0
fonttools==4.43.0
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
lightgbm==4.1.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nbformat==5.1.2
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-aqi==0.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.11.2
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
xgboost==2.0.0
yarg==0.1.9
